Migrate some utilities from UnoDiscardPile to UnoProgressiveRulePack
